% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wb.R
\name{wb_countries}
\alias{wb_countries}
\title{List World Bank countries and regions}
\usage{
wb_countries(per_page = 300)
}
\arguments{
\item{per_page}{Results per page (default 300)}
}
\value{
tibble: iso3, iso2, name, region, income_level, capital,
longitude, latitude
}
\description{
List World Bank countries and regions
}
